rename client file
